requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
tqdm>=4.66.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
umap-learn>=0.5.3
faiss-cpu>=1.7.4
hdbscan>=0.8.33

